Implement caesar cipher
The first project in this section was to implement a casesar cipher.

The shift_letter method works by reducing the letter to a number between
0 and 25, which shows the letter's relative position within the alphabet.
The factor by which the letters are to be shifted is added to that number,
and the number is brought back to a number between 0 and 25 using
modular arithmetic.